+++ src/commands/stats.ts
@@ -6,6 +6,8 @@ import { type FunctionViewStats } from '../../types'
 import { getStopWords } from '../utils/stopwords/stopwords.js'
 import { getLang, getStopwordsLang, setStopwordsLang } from '../utils/lang/lang.js'
 import { tLogsStats } from '../i18n.js'
+import { getWordLengthBuckets, printWordLengthDistribution } from '../feature/stats/wordLengthStats'
+import { viewFileSize } from '../feature/stats/viewFileSize'
 
 /**
  * Gets the current application language.
@@ -17,19 +19,24 @@ const currentLang = getLang()
  * Command to display statistics about a text file, such as line count, word count,
  * unique words, most common words, and stopwords.
  *
- * @param params - Object with the following properties:
- *   - file: Path to the file to analyze.
- *   - exclude: (Optional) Comma-separated words to exclude from the analysis.
- *   - all: (Optional) If null, stopwords and excluded words are filtered out.
- *   - lang: (Optional) Language code ('en' or 'es') for stopwords.
- *   - stopwords: (Optional) If not null, prints the stopwords list for the selected language.
- *   - top: (Optional) Number of top frequent words to display.
- * @returns A promise that resolves when the statistics have been displayed.
+ * @function commandStats
+ * @async
+ * @param {Object} params - Parameters for the stats command.
+ * @param {string} params.file - Path to the file to analyze.
+ * @param {string} [params.exclude] - (Optional) Comma-separated words to exclude from the analysis.
+ * @param {null|boolean} [params.all] - (Optional) If null, stopwords and excluded words are filtered out.
+ * @param {string} [params.lang] - (Optional) Language code ('en' or 'es') for stopwords.
+ * @param {null|boolean} [params.stopwords] - (Optional) If not null, prints the stopwords list for the selected language.
+ * @param {number} [params.top] - (Optional) Number of top frequent words to display.
+ * @returns {Promise<void>} A promise that resolves when the statistics have been displayed.
  *
  * @example
- * ```ts
+ * // Analyze file.txt, exclude "the" and "and", show top 10 words in English
  * await commandStats({ file: './file.txt', exclude: 'the,and', all: null, lang: 'en', stopwords: null, top: 10 });
- * ```
+ *
+ * @example
+ * // Show stopwords for Spanish
+ * await commandStats({ file: './file.txt', lang: 'es', stopwords: true });
  */
 export const commandStats: FunctionViewStats = async ({ file, exclude, all, lang, stopwords, top }) => {
   try {
@@ -87,24 +94,45 @@ export const commandStats: FunctionViewStats = async ({ file, exclude, all, lang
     // Sort and get the top N words
     const sorted = [...wordMap.entries()]
       .sort((a, b) => b[1] - a[1])
-      .slice(0, (top === null) ? 5 : top)
+      .slice(0, top ?? 5)
 
     const mostCommon = sorted[0]
 
+    const buckets = getWordLengthBuckets(words)
+    const totalLinesPorc = ((uniqueWords / totalWords) * 100).toFixed(2)
+    const totalWordsPerLine = (totalWords / totalLines).toFixed(2)
+
     // Output statistics
+    // Shows the stopwords language being used for filtering
     console.log(`${tLogsStats('LOG_STOPWORDS_LANG', currentLang)} ${chalk.green.bold(stopwordsLanguage)}`)
+
+    // Prints the header indicating that statistics for the file are being displayed
     console.log(chalk.bold(`📊 ${tLogsStats('LOG_STATS', currentLang)} ${chalk.underline(file)}:`))
+
+    // Prints a visual separator for clarity
     console.log('──────────────────────────────')
+
+    // Displays the total number of lines in the file
     console.log(`📝 ${tLogsStats('LOG_LINES', currentLang)} ${chalk.green(totalLines)}`)
-    console.log(`🔤 ${tLogsStats('LOG_WORDS', currentLang)}: ${chalk.green(totalWords)}`)
-    console.log(`🔁 ${tLogsStats('LOG_WORDS_UNIQUE', currentLang)}: ${chalk.green(uniqueWords)}`)
+
+    // Displays the total number of words in the file
+    console.log(`🔤 ${tLogsStats('LOG_WORDS', currentLang)}: ${chalk.green(totalWords)} -  ${tLogsStats('LOG_WORDS_PER_LINE_AVG', currentLang)}: ${chalk.underline.bold(totalWordsPerLine)}`)
+
+    // Displays the number of unique words in the file
+    console.log(`🔁 ${tLogsStats('LOG_WORDS_UNIQUE', currentLang)}: ${chalk.green(uniqueWords)} - (${(totalLinesPorc)}%)`)
 
     if (mostCommon != null) console.log(`🏆 ${tLogsStats('LOG_MOST_COMMON_WORD', currentLang)}: ${chalk.yellow(`"${mostCommon[0]}"`)} (${mostCommon[1]} ${tLogsStats('LOG_TIMES', currentLang)})`)
 
-    console.log(`🔝 ${tLogsStats('LOG_TOP', currentLang)} ${(top === null) ? '5' : top} ${tLogsStats('LOG_TOP_WORDS', currentLang)}:`)
+    console.log(`🔝 ${tLogsStats('LOG_TOP', currentLang)} ${top ?? '5'} ${tLogsStats('LOG_TOP_WORDS', currentLang)}:`)
     sorted.forEach(([word, count], i) => {
       console.log(`   ${i + 1}. ${chalk.cyan(word)} ${count}`)
     })
+
+    const printDiagr = printWordLengthDistribution(buckets)
+    console.log(printDiagr)
+
+    const fileSize = await viewFileSize(file)
+    console.log(`\n👀 ${fileSize}`)
   } catch (err) {
     const error = err as Error
     console.error(chalk.red(`⚠️ ${tLogsStats('LOG_ERRROR_READ_FILE', currentLang)}: ${file}`))