+++ src/translation/commands.t/commandStats.t.ts
@@ -9,7 +9,9 @@ export const translationsLogsFileStats = {
     LOG_TIMES: 'veces',
     LOG_TOP: 'Top',
     LOG_TOP_WORDS: 'palabras',
-    LOG_ERRROR_READ_FILE: 'Error al leer el archivo'
+    LOG_ERRROR_READ_FILE: 'Error al leer el archivo',
+    LOG_WORD_LENGTH_DISTRIBUTION: 'Distribución de la longitud de las palabras',
+    LOG_WORDS_PER_LINE_AVG: 'Prom. palabras/línea'
   },
   en: {
     LOG_STOPWORDS_LANG: 'Stopwords language',
@@ -21,6 +23,8 @@ export const translationsLogsFileStats = {
     LOG_TIMES: 'times',
     LOG_TOP: 'Top',
     LOG_TOP_WORDS: 'words',
-    LOG_ERRROR_READ_FILE: 'Error reading the file'
+    LOG_ERRROR_READ_FILE: 'Error reading the file',
+    LOG_WORD_LENGTH_DISTRIBUTION: 'Word length distribution',
+    LOG_WORDS_PER_LINE_AVG: 'Avg. words/line'
   }
 }